Remove sqs dt modules & moved dt logic to original modules
