+++ instrumentation/aws-java-sdk-sqs-1.10.44/build.gradle
@@ -8,7 +8,7 @@ dependencies {
 
 
 jar {
-    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.aws-java-sdk-sqs-1.10.44',
+    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.aws-java-sdk-sqs-spans-1.10.44', 'Enabled': 'false',
             'Implementation-Title-Alias': 'aws-java-sdk-sqs'  }
 }
