+++ instrumentation/aws-java-sdk-sqs-2.1.0/build.gradle
@@ -8,7 +8,7 @@ dependencies {
 
 
 jar {
-    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.aws-java-sdk-sqs-2.1.0',
+    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.aws-java-sdk-sqs-spans-2.1.0', 'Enabled': 'false',
             'Implementation-Title-Alias': 'aws-java-sdk-sqs'  }
 }
