+++ instrumentation/aws-java-sdk-sqs-2.1.0/src/main/java/com/newrelic/utils/SqsV2Util.java
@@ -18,6 +18,7 @@ public class SqsV2Util {
 
     public static final String LIBRARY = "SQS";
     public static final String OTEL_LIBRARY = "aws_sqs";
+    public static final String[] DT_HEADERS = new String[] {"newrelic","NEWRELIC","NewRelic","tracestate","TraceState","TRACESTATE"};
 
     public static MessageProduceParameters generateExternalProduceMetrics(String queueUrl) {
         DestinationData destinationData = DestinationData.parse(queueUrl);
