--- instrumentation/aws-java-sdk-sqs-spans-1.10.44/build.gradle
@@ -1,29 +1,0 @@
-dependencies {
-    implementation(project(":agent-bridge"))
-    implementation("com.amazonaws:aws-java-sdk-sqs:1.10.44")
-    testImplementation("org.apache.httpcomponents:httpclient:4.5.10")
-    testImplementation("org.elasticmq:elasticmq-rest-sqs_2.13:0.15.3")
-}
-
-
-
-jar {
-    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.aws-java-sdk-sqs-spans-1.10.44', 'Enabled': 'false',
-            'Implementation-Title-Alias': 'aws-java-sdk-sqs-spans'  }
-}
-
-verifyInstrumentation {
-    // this instrumentation matches the original sdk starting at 1.3.22 and all the way up to the current version
-    // we only check up to 1.10.44 because it's the same as the sqs dependency onwards
-    // note that in version 1.10.43 and earlier the method Future<SendMessageResult> sendMessageAsync(SendMessageRequest)
-    // does not call the instrumented method Future<SendMessageResult> sendMessageAsync(SendMessageRequest, AsyncHandler<SendMessageRequest, SendMessageResult>)
-    // so that method is not supported until version 1.10.44 and greater.
-    passes 'com.amazonaws:aws-java-sdk:[1.3.22,1.10.44)'
-    exclude 'com.amazonaws:aws-java-sdk:1.6.9' // doesn't exist in maven
-    passes 'com.amazonaws:aws-java-sdk-sqs:[1.10.44,)'
-}
-
-site {
-    title 'AWS SQS'
-    type 'Messaging'
-}
