--- instrumentation/aws-java-sdk-sqs-spans-2.1.0/build.gradle
@@ -1,22 +1,0 @@
-dependencies {
-    implementation(project(":agent-bridge"))
-    implementation("software.amazon.awssdk:sqs:2.1.0")
-    testImplementation("org.apache.httpcomponents:httpclient:4.5.10")
-    testImplementation("org.elasticmq:elasticmq-rest-sqs_2.13:0.15.3")
-}
-
-
-
-jar {
-    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.aws-java-sdk-sqs-spans-2.1.0', 'Enabled': 'false',
-            'Implementation-Title-Alias': 'aws-java-sdk-sqs-spans'  }
-}
-
-verifyInstrumentation {
-    passes 'software.amazon.awssdk:sqs:[2.1.0,)'
-}
-
-site {
-    title 'AWS SQS'
-    type 'Messaging'
-}
