--- instrumentation/aws-java-sdk-sqs-spans-2.1.0/src/test/resources/dt_enabled.yml
@@ -1,3 +1,0 @@
-common: &default_settings
-  distributed_tracing:
-    enabled: true
