+++ settings.gradle
@@ -81,8 +81,6 @@ include 'instrumentation:aws-java-sdk-lambda-1.11.280'
 include 'instrumentation:aws-java-sdk-lambda-2.1'
 include 'instrumentation:aws-java-sdk-sqs-1.10.44'
 include 'instrumentation:aws-java-sdk-sqs-2.1.0'
-include 'instrumentation:aws-java-sdk-sqs-spans-1.10.44'
-include 'instrumentation:aws-java-sdk-sqs-spans-2.1.0'
 include 'instrumentation:aws-java-sdk-s3-1.2.13'
 include 'instrumentation:aws-java-sdk-s3-2.0'
 include 'instrumentation:aws-java-sdk-sns-1.11.12'
